♻️ 663T4T blueprints: split catalog cache helpers
