+++ packages/agentplane/src/commands/blueprints/catalog.ts
@@ -1,4 +1,4 @@
-import { mkdir, cp, mkdtemp, readFile, rm, writeFile, readdir } from "node:fs/promises";
+import { mkdir, cp, mkdtemp, readFile, rm, writeFile } from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
 
@@ -9,10 +9,11 @@ import { validateProjectBlueprintFile } from "../../blueprints/index.js";
 import { CliError, ValidationError } from "../../shared/errors.js";
 import { isRecord } from "../../shared/guards.js";
 import { writeJsonStableIfChanged } from "../../shared/write-if-changed.js";
+import { cacheBlueprintPackage, pickLatestVersion, resolvePackageRoot } from "./catalog-cache.js";
 
 export type CatalogKind = "blueprint" | "pack";
 
-type CatalogEntry = {
+export type CatalogEntry = {
   id: string;
   path?: string;
   name?: string;
@@ -50,7 +51,7 @@ type BlueprintCatalogSource = {
   source: string;
 };
 
-type CatalogBlueprintManifest = {
+export type CatalogBlueprintManifest = {
   schema_version: 1;
   id: string;
   version: string;
@@ -97,10 +98,6 @@ function agentplaneHome(): string {
 
 function blueprintsHomeDir(): string {
   return path.join(agentplaneHome(), "blueprints");
-}
-
-function blueprintCatalogHomeDir(): string {
-  return path.join(agentplaneHome(), "blueprint-catalog");
 }
 
 export function cachedIndexPath(): string {
@@ -424,72 +421,6 @@ function parsePackManifest(raw: unknown): CatalogPackManifest {
   return pack;
 }
 
-function pickLatestVersion(entry: CatalogEntry): NonNullable<CatalogEntry["versions"]>[number] {
-  const versions = entry.versions ?? [];
-  if (versions.length === 0) {
-    throw new ValidationError({ message: `Blueprint ${entry.id} has no installable versions.` });
-  }
-  return versions
-    .toSorted((a, b) => a.version.localeCompare(b.version, undefined, { numeric: true }))
-    .at(-1)!;
-}
-
-async function resolvePackageRoot(extractedDir: string): Promise<string> {
-  const rootManifest = path.join(extractedDir, "blueprint.json");
-  try {
-    await readFile(rootManifest, "utf8");
-    return extractedDir;
-  } catch (err) {
-    const code = (err as { code?: string } | null)?.code;
-    if (code !== "ENOENT") throw err;
-  }
-  const entries = await readdir(extractedDir, { withFileTypes: true });
-  const dirs = entries.filter((entry) => entry.isDirectory()).map((entry) => entry.name);
-  if (dirs.length !== 1) {
-    throw new ValidationError({
-      message:
-        "Blueprint package archive must contain blueprint.json at root or exactly one root directory.",
-    });
-  }
-  const candidate = path.join(extractedDir, dirs[0]);
-  await readFile(path.join(candidate, "blueprint.json"), "utf8");
-  return candidate;
-}
-
-async function cacheBlueprintPackage(opts: {
-  manifest: CatalogBlueprintManifest;
-  manifestSource: string;
-  definitionText: string;
-}): Promise<string> {
-  assertSafeCatalogPathSegment(opts.manifest.version, "Catalog blueprint manifest version");
-  const catalogTarget = path.join(
-    blueprintCatalogHomeDir(),
-    opts.manifest.id,
-    opts.manifest.version,
-  );
-  await rm(catalogTarget, { recursive: true, force: true });
-  await mkdir(path.dirname(catalogTarget), { recursive: true });
-
-  if (!isHttpUrl(opts.manifestSource)) {
-    await cp(path.dirname(opts.manifestSource), catalogTarget, { recursive: true });
-    return catalogTarget;
-  }
-
-  await mkdir(catalogTarget, { recursive: true });
-  await writeJsonStableIfChanged(path.join(catalogTarget, "blueprint.json"), opts.manifest);
-  const definitionTarget = path.resolve(catalogTarget, opts.manifest.definition.path);
-  const catalogRoot = `${catalogTarget}${path.sep}`;
-  if (!definitionTarget.startsWith(catalogRoot)) {
-    throw new ValidationError({
-      message: "Catalog blueprint definition.path must stay within the blueprint package.",
-      context: { path: opts.manifest.definition.path },
-    });
-  }
-  await mkdir(path.dirname(definitionTarget), { recursive: true });
-  await writeFile(definitionTarget, opts.definitionText, "utf8");
-  return catalogTarget;
-}
-
 export async function loadBlueprintManifest(opts: {
   catalogSource: string;
   entry: CatalogEntry;
@@ -601,7 +532,12 @@ export async function installBlueprint(opts: {
     assertSafeCatalogPathSegment(manifest.id, "Catalog blueprint manifest id");
     const definitionSource = resolveNearSource(manifestSource, manifest.definition.path);
     const definitionText = await readText(definitionSource);
-    const cachePath = await cacheBlueprintPackage({ manifest, manifestSource, definitionText });
+    const cachePath = await cacheBlueprintPackage({
+      agentplaneHome: agentplaneHome(),
+      manifest,
+      manifestSource,
+      definitionText,
+    });
     const targetDir = path.join(opts.projectRoot, ".agentplane", "blueprints");
     const targetPath = path.join(targetDir, `${manifest.definition.id}.json`);
     await mkdir(targetDir, { recursive: true });
